refactor: data util 패키지 만들기
