+++ calculator/src/main/java/edu/nextstep/camp/calculator/CalculatorViewModelFactory.kt
@@ -3,7 +3,7 @@ package edu.nextstep.camp.calculator
 import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.github.dodobest.data.Injector
+import com.github.dodobest.data.util.Injector
 import com.github.dodobest.domain.Calculator
 import com.github.dodobest.domain.Expression
 import com.github.dodobest.domain.usecase.AddMemoryUseCase
