+++ data/src/main/java/com/github/dodobest/data/util/ResultRecordExtension.kt
@@ -1,5 +1,6 @@
-package com.github.dodobest.data
+package com.github.dodobest.data.util
 
+import com.github.dodobest.data.ResultRecordEntity
 import com.github.dodobest.domain.ResultRecord
 
 fun ResultRecord.toEntity() : ResultRecordEntity = ResultRecordEntity(this.expression, this.result)
