+++ data/src/main/java/com/github/dodobest/data/util/TestUtil.kt
@@ -1,4 +1,6 @@
-package com.github.dodobest.data
+package com.github.dodobest.data.util
+
+import com.github.dodobest.data.ResultRecordEntity
 
 object TestUtil {
     fun createEmptyResultRecord(): ResultRecordEntity = ResultRecordEntity(
